requirements.txt
bs4==0.0.2
jpype1==1.5.0
pandas==2.2.2
pre-commit==3.7.1
pypdf==4.2.0
requests==2.32.3
tabula-py==2.9.3
